% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipses_pts.R
\name{ellipses_pts}
\alias{ellipses_pts}
\title{Get points of the boundary of an ellipse from multivariate environment data}
\usage{
ellipses_pts(mu, S, alpha = 0.95, n = 100)
}
\arguments{
\item{mu}{A vector of estimated means}

\item{S}{A covariance matrix}

\item{alpha}{the envolving probability boundary}

\item{n}{number of points for the ellipse boundary}
}
\value{
A list with ellipse points for each variable combination
}
\description{
Get points of the boundary of an ellipse from multivariate environment data
}
\examples{
pts <- get_env_var(rawSpOccPnts, stack_1_12_19)
par <- get_ellip_par(pts)
ellipses_pts(par$mu, par$S)
}
