% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_A_matrix.R
\name{get_A_matrix}
\alias{get_A_matrix}
\title{Calculates the inverse of covariance matrix from environmental variables}
\usage{
get_A_matrix(df)
}
\arguments{
\item{df}{A matrix or data frame with environmental variables}
}
\value{
A covariance matrix
}
\description{
Calculates the inverse of covariance matrix from environmental variables
}
\examples{
get_A_matrix(spOccPnts)
}
