% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chi_prob.R
\name{get_chi_prob}
\alias{get_chi_prob}
\title{Chi square probability  distance}
\usage{
get_chi_prob(env, el_pars)
}
\arguments{
\item{env}{A data frame with environmental information according to presence points}

\item{el_pars}{A list with ellipsoid parameters}
}
\value{
A vector with chi square probabilities from Mahalanobis distance
 according to each point in environmental space
}
\description{
Chi square probability  distance
}
\examples{
pars <- get_negloglike_optimr_par(spOccPnts, samMPts)
get_chi_prob(spOccPnts, pars)
}
