% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_density.R
\name{get_density}
\alias{get_density}
\title{Environmental density from presence points}
\usage{
get_density(df, env)
}
\arguments{
\item{df}{A data frame with presence points}

\item{env}{A raster stack with environmental information}
}
\value{
A data frame with environmental information and the probability density for each point
}
\description{
Environmental density from presence points
}
\examples{
get_density(rawSpOccPnts, stack_1_12_19 )
}
