% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ellip_par.R
\name{get_ellip_par}
\alias{get_ellip_par}
\title{Get ellipsoid parameters. A function to compute average and the inverse of covariance matrix from environmental data}
\usage{
get_ellip_par(env)
}
\arguments{
\item{env}{A data frame containing environmental variables}
}
\value{
A list with computed average of environmental variables and the covariance matrix
}
\description{
Get ellipsoid parameters. A function to compute average and the inverse of covariance matrix from environmental data
}
\examples{
get_ellip_par(spOccPnts)
}
