% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ENM_par.R
\name{get_ENM_par}
\alias{get_ENM_par}
\title{get_ENM_par}
\usage{
get_ENM_par(
  occPts,
  env,
  M_shp = NULL,
  method = c("bound", "mahalanobis", "vmmin", "Nelder-Mead")
)
}
\arguments{
\item{occPts}{A data frame with ocurrence points. Has species name, longitude and latitude}

\item{env}{A raster stack}

\item{M_shp}{A shape file as M hypothesis}

\item{method}{Method selection. By default, optimization of the negloglike function with lower bound}
}
\value{
a list of parameters
}
\description{
get ellipsoid parameters for a Ecollogical fundamental niche model.
}
\examples{
get_ENM_par(rawSpOccPnts, stack_1_12, Mshp, method = "bound")
}
