% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_env_var.R
\name{get_env_var}
\alias{get_env_var}
\title{get_env_var
Function to extract environmental values from presence points}
\usage{
get_env_var(df, env)
}
\arguments{
\item{df}{A data frame with presence points}

\item{env}{A raster with environmental layers. Must have a projection string}
}
\description{
get_env_var
Function to extract environmental values from presence points
}
\examples{
get_env_var(rawSpOccPnts, stack_1_12)
}
