% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mahalanobis.R
\name{get_mahalanobis}
\alias{get_mahalanobis}
\title{Mahalanobis distance}
\usage{
get_mahalanobis(df, el_pars)
}
\arguments{
\item{df}{A data frame with environmental information according to presence points}

\item{el_pars}{A list with ellipsoid parameters}
}
\value{
A vector with Mahalanobis distance according to each point in environmental space
}
\description{
Mahalanobis distance
}
\examples{
pars <- get_negloglike_optimr_par(spOccPnts, samMPts)
get_mahalanobis(spOccPnts, pars)
}
