% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_negative_log.R
\name{get_negative_log}
\alias{get_negative_log}
\title{Get the negative log-likelihood value from quadratic forms of environmental information from presence points and a sample from an M hypothesis}
\usage{
get_negative_log(q1, q2)
}
\arguments{
\item{q1}{quadratic terms of presence points}

\item{q2}{quadratic terms of M points}
}
\value{
A single value of the negative log-likelihood
}
\description{
Get the negative log-likelihood value from quadratic forms of environmental information from presence points and a sample from an M hypothesis
}
\examples{
par <- get_optim_par(spOccPnts)
q1 <- mahalanobis(spOccPnts, par$mu, par$A, inverted = TRUE)
q2 <- mahalanobis(samMPts, par$mu, par$A, inverted = TRUE)
get_negative_log(q1, q2)
}
