% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_negloglike_optim_par.R
\name{get_negloglike_optim_par}
\alias{get_negloglike_optim_par}
\title{get_negloglike_optim_par
Function  to optime negloglike function given parameters from presence points and a sample
of environmental points from a M hypothesis. Returns a list of}
\usage{
get_negloglike_optim_par(env_pts, M_pts)
}
\arguments{
\item{env_pts}{A dataframe with environmental variables extracted from presence points}

\item{M_pts}{A dataframe with a sample of environmental values inside an M hypothesis region.}
}
\value{
A list with optimized parameters for negloglike function. mu is a vector of
centroids and S is a covariance matrix.
}
\description{
get_negloglike_optim_par
Function  to optime negloglike function given parameters from presence points and a sample
of environmental points from a M hypothesis. Returns a list of
}
\examples{
MPts <- sam_polyM(Mshp, stack_1_12, 1000)
get_negloglike_optim_par(head(spOccPnts, 30), MPts)

}
