% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_negloglike_optimr_par.R
\name{get_negloglike_optimr_par}
\alias{get_negloglike_optimr_par}
\title{get_negloglike_optim_par
Function  to optime negloglike function given parameters from presence points and a sample
of environmental points from a M hypothesis. Internally runs Returns a list of parameters}
\usage{
get_negloglike_optimr_par(env_pts, M_pts, lower = FALSE)
}
\arguments{
\item{env_pts}{A dataframe with environmental variables extracted from presence points}

\item{M_pts}{A dataframe with a sample of environmental values inside an M hypothesis region.}

\item{lower}{logical if is set create a low bowndarie for parameters from M points}
}
\value{
A list with optimized parameters for negloglike function
}
\description{
get_negloglike_optim_par
Function  to optime negloglike function given parameters from presence points and a sample
of environmental points from a M hypothesis. Internally runs Returns a list of parameters
}
\examples{
get_negloglike_optimr_par(spOccPnts, samMPts)
}
