% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_optim_par.R
\name{get_optim_par}
\alias{get_optim_par}
\title{get_optim_par G
et the parameters for the optimization of
the Mahalanobis ellipse}
\usage{
get_optim_par(df)
}
\arguments{
\item{df}{A data frame with environmental information}
}
\value{
A list with two objects. A vector with centers of ellipse and a
matrix with the inverse of covariance matrix
}
\description{
get_optim_par G
et the parameters for the optimization of
the Mahalanobis ellipse
}
\examples{
get_optim_par(spOccPnts)
}
