% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_realized_centroid.R
\name{get_realized_centroid}
\alias{get_realized_centroid}
\title{Environmental centroid
Get environmental centroid according to raw points}
\usage{
get_realized_centroid(df, env)
}
\arguments{
\item{df}{presence points from species}

\item{env}{A raster stack of environmental information}
}
\value{
A vector with the centroid coordinate from environmental information
}
\description{
Environmental centroid
Get environmental centroid according to raw points
}
\examples{
get_realized_centroid(rawSpOccPnts, stack_1_12_19 )
}
