% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_suitability.R
\name{get_suitability}
\alias{get_suitability}
\title{get_suitability
Function that calculates the log(suitability)}
\usage{
get_suitability(df, el_pars)
}
\arguments{
\item{df}{A data frame with environmental information}

\item{el_pars}{A list with ellipse parameter. Contains a vector of means and
a covariance matrix. Could came from a MLE after optimization or calculated
 from only presence data.}
}
\value{
A vector with suitability
}
\description{
get_suitability
Function that calculates the log(suitability)
}
\examples{
pars <- get_negloglike_optimr_par(spOccPnts, samMPts)
get_suitability(spOccPnts, pars)
}
