% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_latlon.R
\name{guess_latlon}
\alias{guess_latlon}
\title{guess location}
\usage{
guess_latlon(x, lat = NULL, lon = NULL)
}
\arguments{
\item{x}{a data.frame with species records and geographical coordinates.}

\item{lat}{column with latitude information}

\item{lon}{column with longitude information}
}
\value{
A data frame with geographical coordinates
}
\description{
code adapted from the scrubr package - source at github.com/ropensci/scrubr/
original from leaflet package - source at github.com/rstudio/leaflet
}
\examples{
guess_latlon(tidyniche::rawSpOccPnts)
}
