% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialValues.R
\docType{data}
\name{initialValues}
\alias{initialValues}
\title{Initial parameters to estimate negative log likelihood
from Abeillia abeillei presence points. This is a
hummingbird example
A vector with estimated parameters from presence points
contains 2 mu parameters and 4 covariances between variables}
\format{
An object of class \code{numeric} of length 5.
}
\source{
\url{https://doi.org/10.1016/j.ecolmodel.2021.109823/}
}
\usage{
initialValues
}
\description{
Initial parameters to estimate negative log likelihood
from Abeillia abeillei presence points. This is a
hummingbird example
A vector with estimated parameters from presence points
contains 2 mu parameters and 4 covariances between variables
}
\keyword{datasets}
