% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mshp.R
\docType{data}
\name{Mshp}
\alias{Mshp}
\title{Shapefile of M hypothesis for Abeillia abeillei presence points. This is a
hummingbird example.
A SpatialPolygonsDataFrame with the polygon}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 1 rows and 1 columns.
}
\source{
\url{https://doi.org/10.1016/j.ecolmodel.2021.109823/}
}
\usage{
Mshp
}
\description{
Shapefile of M hypothesis for Abeillia abeillei presence points. This is a
hummingbird example.
A SpatialPolygonsDataFrame with the polygon
}
\keyword{datasets}
