% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negloglike.R
\name{negloglike}
\alias{negloglike}
\title{Negative log likelyhood}
\usage{
negloglike(guess, sam1, sam2)
}
\arguments{
\item{guess}{A vector of length 5 when d=2, it contains the mu and A values as elements}

\item{sam1}{A matrix containing the original sample of environmental combinations that correspond to presences}

\item{sam2}{matrix containing a second random sample of environmental combinations which come from the area of study (M)}
}
\value{
A negative log likelyhood value
}
\description{
Negative log likelyhood
}
\examples{
negloglike(initialValues, spOccPnts, samMPts)
}
