% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negloglike_multivariable.R
\name{negloglike_multivariable}
\alias{negloglike_multivariable}
\title{Negative log likelyhood}
\usage{
negloglike_multivariable(guess, sam1, sam2)
}
\arguments{
\item{guess}{A list of length 2, whit two objets, a vector mu of parameters, And a matrix A as the inverse of covariance matrix}

\item{sam1}{A matrix containing the original sample of environmental combinations that correspond to presences}

\item{sam2}{matrix containing a second random sample of environmental combinations which come from the area of study (M)}
}
\value{
A negative log likelyhood value
}
\description{
Negative log likelyhood
}
\examples{
par <- get_ellip_par(spOccPnts)
negloglike_multivariable(par, spOccPnts, samMPts)
}
