% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quad.R
\name{quad}
\alias{quad}
\title{quad  Function that calculates quadratic terms}
\usage{
quad(xi, mu, A)
}
\arguments{
\item{xi}{An environmental data frame}

\item{mu}{A vector with the average of environmental variables}

\item{A}{matrix of inverse of covariance of environmental variables}
}
\value{
This function return a matrix with the quadratic form
}
\description{
quad  Function that calculates quadratic terms
}
\examples{
par <- get_optim_par(spOccPnts)
quad(xi = spOccPnts, mu = par$mu, A = par$A)
}
