% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawSpOccPnts.R
\docType{data}
\name{rawSpOccPnts}
\alias{rawSpOccPnts}
\title{Raw species ocurrence points from Abeillia abeillei presence points after download
from GBIF and cleansing. This is a hummingbird example.
A dataset with three variables. Contains scientific name, longitude and latitude
#' @format A data frame with 73 rows and 3 variables:
\describe{
  \item{species}{Species name}
  \item{long}{Decimal longitide geographical cooordinate, in degrees}
  \item{lat}{Decimal latitude geographical cooordinate, in degrees}
}}
\format{
An object of class \code{data.frame} with 73 rows and 3 columns.
}
\source{
\url{https://doi.org/10.15468/dl.c2ycac}
}
\usage{
rawSpOccPnts
}
\description{
Raw species ocurrence points from Abeillia abeillei presence points after download
from GBIF and cleansing. This is a hummingbird example.
A dataset with three variables. Contains scientific name, longitude and latitude
#' @format A data frame with 73 rows and 3 variables:
\describe{
  \item{species}{Species name}
  \item{long}{Decimal longitide geographical cooordinate, in degrees}
  \item{lat}{Decimal latitude geographical cooordinate, in degrees}
}
}
\keyword{datasets}
