% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samMPts.R
\docType{data}
\name{samMPts}
\alias{samMPts}
\title{Samples points from M hypothesis  to estimate negative log likelihood
from Abeillia abeillei presence points. This is a
hummingbird example.
A dataset with two variables containing points
contains 2 bioclimatic variables
#' @format A data frame with 10000 rows and 2 variables:
\describe{
  \item{bio1WH}{temperature, in C}
  \item{bio12WH}{precipitaion, in mm}
}}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 10000 rows and 2 columns.
}
\source{
\url{https://doi.org/10.1016/j.ecolmodel.2021.109823/}
}
\usage{
samMPts
}
\description{
Samples points from M hypothesis  to estimate negative log likelihood
from Abeillia abeillei presence points. This is a
hummingbird example.
A dataset with two variables containing points
contains 2 bioclimatic variables
#' @format A data frame with 10000 rows and 2 variables:
\describe{
  \item{bio1WH}{temperature, in C}
  \item{bio12WH}{precipitaion, in mm}
}
}
\keyword{datasets}
