% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_1_12_19.R
\docType{data}
\name{stack_1_12_19}
\alias{stack_1_12_19}
\title{A raster stack with bioclimatic variables 1 and 12 from world clime .}
\format{
An object of class \code{RasterStack} of dimension 1080 x 2160 x 3.
}
\source{
\url{https://doi.org/10.1002/joc.5086}
}
\usage{
stack_1_12_19
}
\description{
A raster stack with bioclimatic variables 1 and 12 from world clime .
}
\keyword{datasets}
