% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negloglike_multivariable.R
\name{negloglike_multivariable}
\alias{negloglike_multivariable}
\title{Negative log likelyhood}
\usage{
negloglike_multivariable(mu, S, sam1, sam2)
}
\arguments{
\item{mu}{A vector mu of parameters}

\item{S}{The covariance matrix from environmental data frame}

\item{sam1}{A data.frame containing the original sample of environmental combinations that correspond to presences}

\item{sam2}{A data.frame containing a second random sample of environmental combinations which come from the area of study (M)}
}
\value{
A negative log likelihood value
}
\description{
Negative log likelyhood
}
\examples{
par <- get_ellip_par(spOccPnts)
negloglike_multivariable(par$mu, par$S, spOccPnts, samMPts)
}
